+++ src/router/utils.rs
@@ -400,6 +400,15 @@ pub async fn send_asking_with_policy(
   Ok(())
 }
 
+#[cfg(feature = "replicas")]
+async fn sync_cluster_replicas(inner: &Arc<RedisClientInner>, router: &mut Router) -> Result<(), RedisError> {
+  if inner.config.server.is_clustered() {
+    router.sync_cluster().await
+  } else {
+    router.sync_replicas().await
+  }
+}
+
 /// Repeatedly try to sync the cluster state, reconnecting as needed until the max reconnection attempts is reached.
 #[cfg(feature = "replicas")]
 pub async fn sync_replicas_with_policy(inner: &Arc<RedisClientInner>, router: &mut Router) -> Result<(), RedisError> {
@@ -411,7 +420,7 @@ pub async fn sync_replicas_with_policy(inner: &Arc<RedisClientInner>, router: &m
       let _ = inner.wait_with_interrupt(delay).await?;
     }
 
-    if let Err(e) = router.sync_replicas().await {
+    if let Err(e) = sync_cluster_replicas(inner, router).await {
       _warn!(inner, "Error syncing replicas: {:?}", e);
 
       if e.should_not_reconnect() {